Moved global classes Color, ImageFill etc. to globals
